Rename controlPanelConfigs.js to controlPanels/index.js
